BIG-20942 Cleanup shipping estimator, fix reloading breaking js
