+++ assets/js/theme/cart/shipping-estimator.js
@@ -1,5 +1,6 @@
 import $ from 'jquery';
 import stateCountry from '../common/state-country';
+import utils from 'bigcommerce/stencil-utils';
 
 export default class ShippingEstimator {
 
@@ -8,6 +9,7 @@ export default class ShippingEstimator {
 
         this.$state = $('[data-field-type="State"]', this.$element);
         this.bindStateCountryChange();
+        this.bindEstimatorEvents();
     }
 
     bindStateCountryChange() {
@@ -17,6 +19,53 @@ export default class ShippingEstimator {
                 alert(err);
                 throw new Error(err);
             }
+        });
+    }
+
+    bindEstimatorEvents() {
+        let $estimatorContainer = $('.shipping-estimator'),
+            $estimatorForm = $('.estimator-form');
+
+        $estimatorForm.on('submit', (event) => {
+            let params = {
+                country_id: $('[name="shipping-country"]', $estimatorForm).val(),
+                state_id: $('[name="shipping-state"]', $estimatorForm).val(),
+                zip_code: $('[name="shipping-zip"]', $estimatorForm).val()
+            };
+
+            event.preventDefault();
+
+            utils.api.cart.getShippingQuotes(params, 'cart/shipping-quotes', (err, response) => {
+                $('.shipping-quotes').html(response.content);
+
+                // bind the select button
+                $('.select-shipping-quote').on('click', (event) => {
+                    let quoteId = $('.shipping-quote:checked').val();
+
+                    event.preventDefault();
+
+                    utils.api.cart.submitShippingQuote(quoteId, (response) => {
+                        location.reload();
+                    });
+                });
+            });
+        });
+
+        $('.shipping-estimate-show').on('click', (event) => {
+            event.preventDefault();
+
+            $(event.currentTarget).hide();
+            $estimatorContainer.show();
+            $('.shipping-estimate-hide').show();
+        });
+
+
+        $('.shipping-estimate-hide').on('click', (event) => {
+            event.preventDefault();
+
+            $estimatorContainer.hide();
+            $('.shipping-estimate-show').show();
+            $('.shipping-estimate-hide').hide();
         });
     }
 }
